feat: added validation

+++ Routes/product.js
@@ -1,5 +1,11 @@
 const ProductRouter = require('express').Router();
 const Product = require('../Models/product');
+const {
+  validateStringField,
+  validatePrice,
+  validateQuantity,
+} = require('../Validators');
+const { isEmpty } = require('../Utilities');
 
 // Get All Products
 ProductRouter.get('/', async (req, res) => {
@@ -27,6 +33,33 @@ ProductRouter.get('/:id', async (req, res) => {
 ProductRouter.post('/', async (req, res) => {
   const { name, description, price, quantity } = req.body;
 
+  const errorMessages = [];
+
+  const nameValidation = validateStringField(name, 'name');
+  if (nameValidation.isInValid) {
+    errorMessages.push(nameValidation.errorMessage);
+  }
+
+  const descriptionValidation = validateStringField(description, 'description');
+  if (descriptionValidation.isInValid) {
+    errorMessages.push(descriptionValidation.errorMessage);
+  }
+
+  const priceValidation = validatePrice(price);
+  if (priceValidation.isInvalid) {
+    errorMessages.push(priceValidation.errorMessage);
+  }
+
+  const quantityValidation = validateQuantity(quantity);
+  if (quantityValidation.isInvalid) {
+    errorMessages.push(quantityValidation.errorMessage);
+  }
+
+  if (errorMessages.length > 0) {
+    res.status(400).json({ errors: errorMessages });
+    return;
+  }
+
   const productEntity = new Product({
     name: name,
     description: description,
@@ -45,10 +78,49 @@ ProductRouter.post('/', async (req, res) => {
 // Update Product
 ProductRouter.put('/:id', async (req, res) => {
   const id = req.params.id;
+
+  if (isEmpty(id)) {
+    res.status(400).json({ error: 'Id is not provided' });
+    return;
+  }
+
   const { name, description, price, quantity } = req.body;
+
+  const errorMessages = [];
+
+  const nameValidation = validateStringField(name, 'name');
+  if (nameValidation.isInValid) {
+    errorMessages.push(nameValidation.errorMessage);
+  }
+
+  const descriptionValidation = validateStringField(description, 'description');
+  if (descriptionValidation.isInValid) {
+    errorMessages.push(descriptionValidation.errorMessage);
+  }
+
+  const priceValidation = validatePrice(price);
+  if (priceValidation.isInvalid) {
+    errorMessages.push(priceValidation.errorMessage);
+  }
+
+  const quantityValidation = validateQuantity(quantity);
+  if (quantityValidation.isInvalid) {
+    errorMessages.push(quantityValidation.errorMessage);
+  }
+
+  if (errorMessages.length > 0) {
+    res.status(400).json({ errors: errorMessages });
+    return;
+  }
 
   try {
     const product = await Product.findById(id);
+
+    if(!product) {
+      res.status(400).json({error: 'Product not found'});
+      return;
+    }
+
     product.name = name;
     product.description = description;
     product.price = price;
@@ -66,8 +138,19 @@ ProductRouter.put('/:id', async (req, res) => {
 ProductRouter.delete('/:id', async (req, res) => {
   const id = req.params.id;
 
+  if(isEmpty(id)) {
+    res.status(400).json({error: 'Id is not provided'});
+    return;
+  }
+
   try {
     const product = await Product.findById(id);
+
+    if(!product) {
+      res.status(400).json({error: 'Product not found'});
+      return;
+    }
+
     await product.deleteOne();
     res.json(product);
   } catch (error) {
